requirements.txt
langgraph
langchain[openai]
python-dotenv
Flask